Update Roots of  Quadratic
Finds the roots of a quadratic, given a function, which can be written in standard or vertex form. It can have the function on one side of the equation, or on both, with both sides not needing matching forms. Cubics can also be solved if they are in standard form and have no constant term